+++ scripts/extractLuauDocs.py
@@ -1,7 +1,5 @@
 from json import load, dump
 
-PREFIX_LENGTH = 7
-
 with open("api-docs/en-us.json") as f:
     data = load(f)
 
@@ -9,59 +7,63 @@ luau_docs = {k: v for k, v in data.items() if k.startswith("@luau")}
 vector_lib_docs = {k: v for k, v in data.items() if k.startswith("@roblox/global/vector")}
 new_vector_lib_docs = {}
 
-# Changes docs entries for the vector library to be consistent with other entries.
+# changes docs keys and entries for the vector library to be consistent with other entries.
 for docs_key, docs_entry in vector_lib_docs.items():
-    new_vector_lib_docs["@luau" + docs_key[PREFIX_LENGTH:]] = docs_entry
+    new_vector_lib_docs[docs_key.replace("@roblox", "@luau")] = docs_entry
     
     param_val = docs_entry.get("params")
     ret_val = docs_entry.get("returns")
     
+    # no param = no return -> no need to correct anything else.
     if param_val is None:
         continue
     
-    # Correct the params
+    # correct the params to @luau.
     for param_entry in param_val:
-        param_docs = param_entry["documentation"]
-        param_entry["documentation"] = "@luau" + param_docs[PREFIX_LENGTH:]
+        param_entry["documentation"] = param_entry["documentation"].replace("@roblox", "@luau")
     
-    # Correct the return value.
-    val = ret_val[0]
-    ret_val[0] = "@luau" + ret_val[0][PREFIX_LENGTH:]
+    # correct the return to @luau.
+    ret_val[0] = ret_val[0].replace("@roblox", "@luau")
 
-# Correct the keys.
-for k, v in vector_lib_docs["@roblox/global/vector"]["keys"].items():
-    new_vector_lib_docs["@luau/global/vector"]["keys"][k] = "@luau" + v[PREFIX_LENGTH:]
+# correct the docs keys to @luau.
+for key, value in vector_lib_docs["@roblox/global/vector"]["keys"].items():
+    new_vector_lib_docs["@luau/global/vector"]["keys"][key] = value.replace("@roblox", "@luau")
 
 luau_docs.update(new_vector_lib_docs)
 
-# Changes the URL from directing to the roblox docs site to the luau.org site.
+# redirect the URL from the create.roblox.com site to the luau.org site.
 for docs_key, docs_entry in luau_docs.items():
     link = docs_entry.get("learn_more_link")
     
     if link is None:
         continue
     
-    # If the URL is from a built-in library.
+    # if the URL is from a built-in library.
     if link.startswith("https://create.roblox.com/docs/reference/engine/libraries/"):
         
-        # Find the index from the right of the path separator ("/") and the fragment ("#")
+        # find the index from the right of the path separator ("/") and the fragment ("#")
         path_sep_idx = link.rfind("/")
         fragment_idx = link.rfind("#")
         
-        # Corrects the fragment index if the fragment is found.
-        # This is only the case for directing to specific functions.
+        # corrects the fragment index if the fragment is found.
+        # this is only the case for directing to specific functions.
         fragment_idx = None if fragment_idx == -1 else fragment_idx + 1
         
-        # Add the missing fragment symbol ("#") because the roblox docs link doesn't have it,
+        # add the missing fragment symbol ("#") because the roblox docs link doesn't have it,
         # while the luau.org link does.
-        link = link[:path_sep_idx + 1] + "#" + link[path_sep_idx + 1:]
+        link = f"{link[:path_sep_idx + 1]}#{link[path_sep_idx + 1:]}"
         
-        # Add formats the new link.
-        luau_docs[docs_key]["learn_more_link"] = "https://luau.org/library/" + link[path_sep_idx + 1:fragment_idx] + "-library"
+        # formats the new link.
+        luau_docs[docs_key]["learn_more_link"] = f"https://luau.org/library/{link[path_sep_idx + 1:fragment_idx]}-library"
         
-    # Otherwise, it's from a global function.
+    # otherwise, it's from a global function. cannot do more than this.
     else:
         luau_docs[docs_key]["learn_more_link"] = "https://luau.org/library/#global-functions"
+
+# replacing `require()` description with a general one. this new description is based off of the original description.
+luau_docs["@luau/global/require/param/0"]["documentation"] = "The path to a file or a directory that will be executed to retrieve the return value it provides. Paths may use aliases defined in <code>.config.luau</code> or <code>.luaurc</code>. If the path is a file, executes that file. If the path is a directory, finds an <code>init.luau</code> or <code>init.lua</code> file inside of that directory to execute."
+luau_docs["@luau/global/require/return/0"]["documentation"] = "What the file returned (usually a table or a function)."
+luau_docs["@luau/global/require"]["documentation"] = "Returns the value that was returned by the given path, running it if it has not been run yet."
 
 with open("api-docs/luau-en-us.json", "w") as f:
     dump(luau_docs, f, indent=2)